+++ packages/ui/certd-server/db/migration-pg/v10006__pipeline_title.sql
@@ -2,4 +2,4 @@ alter table pi_pipeline alter column title type varchar(100) using title::varcha
 alter table pi_pipeline alter column content type text using content::text;
 alter table pi_storage alter column value type text using value::text;
 
-alter table pi_pipeline  add "order" integer default 0;
+alter table pi_pipeline  add "order" bigint default 0;
